+++ packages/ai/src/types/index.ts
@@ -1,10 +1,7 @@
 export type { JSONSchema7 } from '@ai-sdk/provider';
 export type { Embedding, EmbeddingModel } from './embedding-model';
-export type {
-  ImageModel,
-  ImageGenerationWarning as ImageModelCallWarning,
-  ImageModelProviderMetadata,
-} from './image-model';
+export type { EmbeddingModelMiddleware } from './embedding-model-middleware';
+export type { ImageModel, ImageModelProviderMetadata } from './image-model';
 export type { ImageModelResponseMetadata } from './image-model-response-metadata';
 export type { JSONValue } from './json-value';
 export type {
@@ -14,21 +11,18 @@ export type {
   ToolChoice,
 } from './language-model';
 export type { LanguageModelMiddleware } from './language-model-middleware';
-export type { EmbeddingModelMiddleware } from './embedding-model-middleware';
 export type { LanguageModelRequestMetadata } from './language-model-request-metadata';
 export type { LanguageModelResponseMetadata } from './language-model-response-metadata';
 export type { Provider } from './provider';
 export type { ProviderMetadata } from './provider-metadata';
 export type { RerankingModel } from './reranking-model';
-export type { SpeechModel, SpeechWarning } from './speech-model';
+export type { SpeechModel } from './speech-model';
 export type { SpeechModelResponseMetadata } from './speech-model-response-metadata';
-export type {
-  TranscriptionModel,
-  TranscriptionWarning,
-} from './transcription-model';
+export type { TranscriptionModel } from './transcription-model';
 export type { TranscriptionModelResponseMetadata } from './transcription-model-response-metadata';
 export type {
   EmbeddingModelUsage,
-  LanguageModelUsage,
   ImageModelUsage,
+  LanguageModelUsage,
 } from './usage';
+export type { Warning } from './warning';