+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ai-sdk/xai
 
+## 3.0.0-beta.45
+
+### Patch Changes
+
+- 457318b: chore(provider,ai): switch to SharedV3Warning and unified warnings
+- Updated dependencies [457318b]
+  - @ai-sdk/openai-compatible@2.0.0-beta.38
+  - @ai-sdk/provider@3.0.0-beta.20
+  - @ai-sdk/provider-utils@4.0.0-beta.37
+
 ## 3.0.0-beta.44
 
 ### Patch Changes